refactor(terraform-module): extract getRegistryRepository to utils
